# WIP: PERF: Cythonize fillna