--- a/pandas/_libs/algos.pyi
+++ b/pandas/_libs/algos.pyi
@@ -127,19 +127,19 @@ def fillna1d(
     arr: np.ndarray,  # fillna_t[:]
     value: object,  # fillna_t
     limit: int,
-    inf_as_na: bool,
+    inf_as_na: bool = False,
 ) -> None: ...
 def fillna1d_multi_values(
     arr: np.ndarray,  # fillna_t[:]
-    values: np.ndarray,  # fillna_values_t
+    value: np.ndarray,  # fillna_values_t
     limit: int,
-    inf_as_na: bool,
+    inf_as_na: bool = False,
 ) -> None: ...
 def fillna2d(
     arr: np.ndarray,  # fillna_t[:]
-    values: object,  # fillna_t
+    value: object,  # fillna_t
     limit: int,
-    inf_as_na: bool,
+    inf_as_na: bool = False,
 ) -> None: ...
 def is_monotonic(
     arr: np.ndarray,  # ndarray[algos_t, ndim=1]