--- a/pandas/_libs/algos.pyx
+++ b/pandas/_libs/algos.pyx
@@ -1,4 +1,3 @@
-# cython: profile=True
 import cython
 from cython import Py_ssize_t