--- a/pandas/core/internals/blocks.py
+++ b/pandas/core/internals/blocks.py
@@ -435,39 +435,63 @@ class Block(PandasObject):
         ObjectBlock and try again
         """
         inplace = validate_bool_kwarg(inplace, "inplace")
-
-        mask = isna(self.values)
-        mask, noop = validate_putmask(self.values, mask)
-
-        if limit is not None:
-            limit = libalgos.validate_limit(None, limit=limit)
-            mask[mask.cumsum(self.ndim - 1) > limit] = False
+        arr = self if inplace else self.copy()
 
         if not self._can_hold_na:
-            if inplace:
-                return [self]
+            return [arr]
+
+        if not self.is_extension:
+            if self._can_hold_element(value):
+                if self.ndim == 1:
+                    libalgos.fillna1d(arr.values, value=value, limit=limit)
+                else:
+                    libalgos.fillna2d(arr.values, value=value, limit=limit)
+                return arr._maybe_downcast([arr], downcast)
+            elif self.ndim == 1 or self.shape[0] == 1:
+                blk = self.coerce_to_target_dtype(value)
+                # bc we have already cast, inplace=True may avoid an extra copy
+                return blk.fillna(value, limit=limit, inplace=True, downcast=None)
             else:
-                return [self.copy()]
-
-        if self._can_hold_element(value):
-            nb = self if inplace else self.copy()
-            putmask_inplace(nb.values, mask, value)
-            return nb._maybe_downcast([nb], downcast)
-
-        if noop:
-            # we can't process the value, but nothing to do
-            return [self] if inplace else [self.copy()]
-
-        elif self.ndim == 1 or self.shape[0] == 1:
-            blk = self.coerce_to_target_dtype(value)
-            # bc we have already cast, inplace=True may avoid an extra copy
-            return blk.fillna(value, limit=limit, inplace=True, downcast=None)
-
+                # operate column-by-column
+                return self.split_and_operate(
+                    type(self).fillna,
+                    value,
+                    limit=limit,
+                    inplace=inplace,
+                    downcast=None,
+                )
         else:
-            # operate column-by-column
-            return self.split_and_operate(
-                type(self).fillna, value, limit=limit, inplace=inplace, downcast=None
-            )
+            # print("hello")
+            return [
+                self.make_block_same_class(
+                    values=self.values.fillna(value=value, limit=limit)
+                )
+            ]
+        # mask = isna(self.values)
+        # mask, noop = validate_putmask(self.values, mask)
+        #
+        # if limit is not None:
+        #     #limit = libalgos.validate_limit(None, limit=limit)
+        #     #mask[mask.cumsum(self.ndim - 1) > limit] = False
+        #
+        # if self._can_hold_element(value):
+        #     putmask_inplace(arr.values, mask, value)
+        #     return arr._maybe_downcast([arr], downcast)
+        #
+        # if noop:
+        #    # we can't process the value, but nothing to do
+        #    return arr
+        #
+        # elif self.ndim == 1 or self.shape[0] == 1:
+        #     blk = self.coerce_to_target_dtype(value)
+        #     # bc we have already cast, inplace=True may avoid an extra copy
+        #     return blk.fillna(value, limit=limit, inplace=True, downcast=None)
+        #
+        # else:
+        #     # operate column-by-column
+        #     return self.split_and_operate(
+        #         type(self).fillna, value, limit=limit, inplace=inplace, downcast=None
+        #     )
 
     @final
     def _split(self) -> list[Block]: