… asv_bench/benchmarks/replace.py