@jbrockmendel 2021-07-09
Good idea, yes.